// lib/types.ts
import { Prisma } from '@prisma/client'
import { Message } from 'ai'

export interface Chat extends Record<string, any> {
  id: string
  title: string
  createdAt: Date
  userId: string
  path: string
  messages: Message[]
  sharePath?: string
}

export type ServerActionResult<Result> = Promise<
  | Result
  | {
      error: string
    }
>

export interface Session {
  user: {
    id: string
    email: string
  }
}

export interface AuthResult {
  type: string
  message: string
}

export type ConnectionWithSenderReceiver = Prisma.ConnectionGetPayload<{
  include: {
    sender: true
    receiver: true
  }
}>